pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "jumper-extension"
version = "0.3.6"
description = "IPython extension for monitoring cell performance"
readme={ file='README.md', content-type="text/markdown"}
license-files=["LICENSE"]
requires-python = ">=3.8"
dependencies = [
    "ipython",
    "matplotlib",
    "plotly",
    "kaleido",
    "pandas",
    "numpy",
    "ipympl",
    "psutil",
    "itables",
    "ipywidgets",
    "jinja2",
    "pydantic>=2.0",
    "pyyaml",
]
urls = {Homepage = "https://github.com/ScaDS/jumper_jupyter_performance"}

[project.entry-points."IPython.extensions"]
jumper_extension = "jumper_extension:load_ipython_extension"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "pynvml",
]

docs = [
    "mkdocs-material",
    "mkdocstrings[python]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["jumper_extension*"]

[tool.setuptools.package-data]
"jumper_extension.config" = ["*.yaml"]
"jumper_extension.templates.report" = ["*.html", "*.css"]
"jumper_extension.templates.visualizer.plotly" = ["**/*"]
"jumper_extension.monitor.backends.native_c" = ["collector.c", "Makefile", "jumper_collector"]
